Added new features
Decided to move away from new form and add a panel on top of the existing form. When the user presses escape the settingsMenu panel comes up with 4 buttons. 1 button is going to be used to display controls, this button does not work at the moment. Button 2 can be used to make the window full screen, this revealed the images do not scale. Pressing button 2 again reverts the window back to a windowed setting. Button 3 quits the game completely. Button 4 can be used to go back to the game, this button has a bug at the moment where if you press it, it will take you back to the game, but keyboard strokes are no longer detected.